refactor: split account API module by responsibility
